requirements.txt
streamlit-1.5.0-py2.py3-none-any.whl